pyproject.toml
[project]
name = "sahool-unified"
version = "16.0.0"
description = "SAHOOL - Smart Agricultural Platform for Yemen & Middle East"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "Proprietary"}
authors = [
    {name = "KAFAAT Team", email = "dev@kafaat.io"}
]

[project.optional-dependencies]
base = [
    "fastapi==0.126.0",
    "uvicorn[standard]==0.27.0",
    "pydantic>=2.10.0,<3.0.0",
    "httpx==0.28.1",
    "python-dotenv==1.0.1",
    "aiohttp>=3.11.12",  # Security fix for CVE-2025-53643
]
infrastructure = [
    "asyncpg==0.30.0",
    "tortoise-orm==0.21.7",
    "aerich==0.7.2",
    "nats-py==2.9.0",
    "redis==5.2.1",
    "python-jose[cryptography]==3.4.0",  # CVE-2024-33663, CVE-2024-33664
    "passlib[bcrypt]==1.7.4",
]
observability = [
    "prometheus-client==0.21.1",
    "opentelemetry-api==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "structlog==24.4.0",
]
testing = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "pytest-xdist==3.5.0",
]
ml = [
    "Pillow==11.0.0",
    "numpy>=1.26.0,<2.1.0",  # Pinned for TensorFlow compatibility
    "tensorflow-cpu==2.18.0",
]
dev = [
    "ruff==0.8.4",
    "pre-commit==4.0.1",
    "black==24.10.0",
    "isort==5.13.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E","F","I","UP","B","SIM","N","W","C4"]
ignore = [
    "B008",  # Depends/Query in FastAPI function defaults is standard pattern
    "B904",  # Exception chaining - not critical, best effort fixed
    "E501",  # Line too long - handled by formatter
    "B007",  # Loop control variable not used - often intentional
    "B006",  # Mutable default argument - sometimes intentional
    "N814",  # CamelCase imported as constant - common pattern
    "N815",  # Variable in class scope mixedCase - existing codebase style
    "N806",  # Variable should be lowercase - false positives
    "N805",  # First arg should be self - sometimes intentional (classmethods)
    "N999",  # Invalid module name - directory structure is intentional
    "N818",  # Exception naming - existing codebase style
    "N802",  # Function name should be lowercase - some patterns are intentional
    "N803",  # Argument name should be lowercase - math variables like X, Y
    "B905",  # zip() without strict - Python 3.10+ only
    "E712",  # Comparison to True/False - SQLAlchemy filter pattern
    "E741",  # Ambiguous variable name - sometimes intentional (l, O, I)
    "E722",  # Bare except - sometimes needed for catch-all
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["E402", "B017"]
"**/__init__.py" = ["F401"]
"**/main.py" = ["E402"]  # Allow imports after sys.path modifications

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q --cov=shared --cov-report=term-missing --cov-report=html:coverage_html"
testpaths = ["tests", "kernel/services"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning:tortoise.*",
    "ignore::DeprecationWarning:pydantic.*",
]
markers = [
    "unit: Unit tests (fast, no I/O)",
    "integration: Integration tests (API, database)",
    "smoke: Smoke tests (import verification)",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["shared", "kernel"]
branch = true
omit = [
    "**/tests/*",
    "**/__pycache__/*",
    "**/migrations/*",
    "**/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 60
show_missing = true
skip_empty = true

[tool.coverage.html]
directory = "coverage_html"
